warning: fix all cast from pointer to integer of different size
Ticket: #185

+++ osaf/services/saf/plmsv/plms/plms_dbg_utils.c
@@ -153,7 +153,7 @@ plms_cb_dump_routine ()
 			else
 				fprintf(fp,"\n Invoke Callback: NULL");
 			fprintf(fp,"\n First PLM notify id:%llu",(unsigned long long)ent->first_plm_ntf_id);
-			fprintf(fp,"\n PLMS Timer id:%d",(int)ent->tmr.timer_id);
+			fprintf(fp,"\n PLMS Timer id:%d",ent->tmr.timer_id);
 			if (ent->tmr.tmr_type == PLMS_TMR_NONE)
 				fprintf(fp,"\n PLMS Timer type:PLMS_TMR_NONE");
 			else if (ent->tmr.tmr_type == PLMS_TMR_EE_INSTANTIATING)
